2.02 works only in k3d cluster, hardcoded url for backend app
A few things: the overall architecture and design of front and back sucks (dist folder built in a different app and served statically,
docker images built separatedly - could use compose with k8s, haven't check that out yet, hardcoded env vars due to Vite being a pain, etc),
it is done like this with the sole purpose of advancing with the course exercises.

Ideally the ingress would work with an nginx reverse proxy, funneling all requests to the exposed url in the host machine.
Variables are a mess with Vite apps in conjunction with k8s, I need to research more into this eventually.

So todo list for when I REALLY need to work with k8s in production envs:
	- Manage variables correctly in vite and k8s (set them in cloud provider, Dockerfile, docker compose, env vars, etc).
	- Understand deeply networking: container<>container, pod<>pod.
	- Set up an nginx reverse proxy
	- Automate docker builds